+++ server/src/helper/all-enums.ts
@@ -9,6 +9,7 @@ import { BombVariant } from '../enum/bomb-variant.enum';
 import { ButtonAction } from '../enum/button-action.enum';
 import { CacheFlag } from '../enum/cache-flag.enum';
 import { Card } from '../enum/card.enum';
+import { Challenge } from '../enum/challenge.enum';
 import { EntityType } from '../enum/entity-type.enum';
 
 export const AllEnums = [
@@ -23,5 +24,6 @@ export const AllEnums = [
   ['ButtonAction', ButtonAction],
   ['CacheFlag', CacheFlag],
   ['Card', Card],
+  ['Challenge', Challenge],
   ['EntityType', EntityType]
 ];