+++ server/src/enum/chest-subtype.enum.ts
@@ -1,0 +1,4 @@
+export enum ChestSubType {
+  CHEST_OPENED = 0,
+  CHEST_CLOSED
+}