+++ client/src/test/completion.test.ts
@@ -26,6 +26,7 @@ suite('Should do completion', () => {
         { label: 'Difficulty', kind: vscode.CompletionItemKind.Enum },
         { label: 'Direction', kind: vscode.CompletionItemKind.Enum },
         { label: 'DoorSlot', kind: vscode.CompletionItemKind.Enum },
+        { label: 'DoorState', kind: vscode.CompletionItemKind.Enum },
         { label: 'EntityType', kind: vscode.CompletionItemKind.Enum }
       ]
     });