+++ client/src/test/completion.test.ts
@@ -27,6 +27,7 @@ suite('Should do completion', () => {
         { label: 'Direction', kind: vscode.CompletionItemKind.Enum },
         { label: 'DoorSlot', kind: vscode.CompletionItemKind.Enum },
         { label: 'DoorState', kind: vscode.CompletionItemKind.Enum },
+        { label: 'DoorVariant', kind: vscode.CompletionItemKind.Enum },
         { label: 'EntityType', kind: vscode.CompletionItemKind.Enum }
       ]
     });