+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Completion for Card Enum
 Completion for Challenge Enum
 Completion for ChampionColor Enum
+Completion for ChestSubType Enum
 ## 0.0.12
 ### Added
 Completion for CacheFlag Enum