+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Completion for Card Enum
 Completion for Challenge Enum
 Completion for ChampionColor Enum
 Completion for ChestSubType Enum
+Completion for CoinSubType Enum
 ## 0.0.12
 ### Added
 Completion for CacheFlag Enum