+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Completion for Challenge Enum
 Completion for ChampionColor Enum
 Completion for ChestSubType Enum
 Completion for CoinSubType Enum
+Completion for Difficulty Enum
 ## 0.0.12
 ### Added
 Completion for CacheFlag Enum