+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Completion for CoinSubType Enum
 Completion for Difficulty Enum
 Completion for Direction Enum
 Completion for DoorSlot Enum
+Completion for DoorState Enum
 ## 0.0.12
 ### Added
 Completion for CacheFlag Enum