+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Change Log](https://keepachangelog.com/en/1.0.0/)
+## 0.0.14
+### Added
+Completion for EntityCollisionClass
 ## 0.0.13
 ### Added
 Completion for Card Enum